// src/shared/types/Querys/IFormulario.ts
export interface IFormulario {
  id: number;
  nombre: string;
  apellido: string;
  correo_electronico: string;
  telefono: string;
  cuil: string;
  dni: string;
  fecha_nacimiento: string;
  nacionalidad: string;
  provincia: string;
  localidad: string;
  domicilio: string;
  codigo_postal: string;
  empresa: string;
  categoria: string;
  objetivo: string;
  created: string;
  lugar_nacimiento: string;
  fecha_ingreso: string;
  estado_civil: string;
  cantidad_hijos: number;
  telefono_empresa: string;
  domicilio_empresa: string;
  localidad_empresa: string;
  cod_postal_empresa: string;
  provincia_empresa: string;
  cuit_empresa: string;
  numero_agencia_empresa: string;
}